Properly detect if sites are missing base URL
Previously, if a site was missing the webring base/index URL, but had
the next/previous URL(s), no error was reported, because the base URL
would be found as a substring of the other URL(s).

This is now fixed by tracking the position at which each match is found,
so a given URL doesn't count twice.