Don't perform site checks when constructing webring
It's a little unintuitive and probably unexpected that checks are
performed while the webring is created. The checking is now removed from
the constructor, and members are initialized to have check_successful
set to true until a check is done.

Also adds a unit test to ensure check logic works.