feat: use tracing instead of log
`tracing` is like `log` but gives structured information about where and
when events happen. It has a compatibility layer for `log`, so libraries
using `log` will still show up in our new logging. This does mean the
log output format changes a bit, but we don't have anything relying on
the structure of our logs, so it shouldn't matter.